rename mixin

+++ src/main/java/gravity_changer/mixin/client/CameraMixin.java
@@ -75,9 +75,9 @@ public abstract class CameraMixin {
     
         Quaternionf gravityRotation = animation.getCurrentGravityRotation(gravityDirection, timeMs);
         
-        double entityX = Mth.lerp((double) tickDelta, focusedEntity.xo, focusedEntity.getX());
-        double entityY = Mth.lerp((double) tickDelta, focusedEntity.yo, focusedEntity.getY());
-        double entityZ = Mth.lerp((double) tickDelta, focusedEntity.zo, focusedEntity.getZ());
+        double entityX = Mth.lerp(tickDelta, focusedEntity.xo, focusedEntity.getX());
+        double entityY = Mth.lerp(tickDelta, focusedEntity.yo, focusedEntity.getY());
+        double entityZ = Mth.lerp(tickDelta, focusedEntity.zo, focusedEntity.getZ());
         
         double currentCameraY = Mth.lerp(tickDelta, this.eyeHeightOld, this.eyeHeight);
